refactor: api 관련 util 함수 파일 분리 - utils/api 폴더 생성 및 내부에 함수 이동 및 적용
