--- frontend/src/api/utils.ts
@@ -1,13 +1,0 @@
-import { ACCESS_TOKEN_KEY } from "@src/@constants/api";
-import { getCookie } from "@src/@utils";
-
-export const getPublicHeaders = () => {
-  return { "Access-Control-Allow-Origin": "*" };
-};
-
-export const getPrivateHeaders = () => {
-  return {
-    "Access-Control-Allow-Origin": "*",
-    authorization: `Bearer ${getCookie(ACCESS_TOKEN_KEY)}`,
-  };
-};
