+++ frontend/src/pages/SpecificDateFeed/index.tsx
@@ -23,7 +23,8 @@ import useSetReminderTargetMessage from "@src/hooks/useSetReminderTargetMessage"
 import useTopScreenEventHandler from "@src/hooks/useTopScreenEventHandlers";
 
 import { FlexColumn } from "@src/@styles/shared";
-import { extractResponseMessages, parseMeridiemTime } from "@src/@utils";
+import { parseMeridiemTime } from "@src/@utils";
+import { extractResponseMessages } from "@src/@utils/api";
 
 function SpecificDateFeed() {
   const { key: queryKey } = useLocation();
